requirements.txt
aiohttp==3.7.3
alembic==1.5.2
astroid==2.4.2
async-timeout==3.0.1
asyncpg==0.21.0
attrs==20.3.0
chardet==3.0.4
gino==1.0.1
idna==3.1
iniconfig==1.1.1
isort==5.7.0
lazy-object-proxy==1.4.3
Mako==1.1.4
MarkupSafe==1.1.1
mccabe==0.6.1
multidict==5.1.0
packaging==20.8
pluggy==0.13.1
py==1.10.0
pylint==2.6.0
pyparsing==2.4.7
pytest==6.2.2
python-dateutil==2.8.1
python-editor==1.0.4
PyYAML==5.4.1
six==1.15.0
SQLAlchemy==1.3.22
toml==0.10.2
typing-extensions==3.7.4.3
ujson==4.0.2
uvloop==0.14.0
wrapt==1.12.1
yarl==1.6.3
